fix helpers bugs
