--- www/src/helpers/helpers.ts
@@ -1,281 +1,0 @@
-import type {Vector} from "./vector.js";
-import {Color} from "./color.js";
-
-/**
- * @param {string} url
- * @param {*} callback
- */
-function loadText(url: string, callback: any) {
-    let xobj = new XMLHttpRequest();
-    xobj.overrideMimeType('application/json');
-    xobj.open('GET', url, true); // Replace 'my_data' with the path to your file
-    xobj.onreadystatechange = function() {
-        if (xobj.readyState == 4 && xobj.status == 200) {
-            // Required use of an anonymous callback as .open will NOT return a value but simply returns undefined in asynchronous mode
-            callback(xobj.responseText);
-        }
-    };
-    xobj.send(null);
-}
-
-export class Helpers {
-    [key: string]: any;
-
-    static cache = new Map();
-    static fetch : Function;
-    static fs;
-    static fetchJSON: (url: any, useCache?: boolean, namespace?: string) => Promise<any>;
-    static fetchBinary: (url: any) => Promise<ArrayBuffer>;
-
-    static setCache(cache) {
-        Helpers.cache = cache;
-    }
-
-    static getCache() : Map<any, any> {
-        return Helpers.cache;
-    }
-
-    //
-    angleTo(pos, target) {
-        let angle = Math.atan2(target.x - pos.x, target.z - pos.z);
-        return (angle > 0) ? angle : angle + 2 * Math.PI;
-    }
-
-    // clamp
-    static clamp(x : number, min : number, max : number) : number {
-        if(!min) {
-            min = 0;
-        }
-        if(!max) {
-            max = 1;
-        }
-        if(x < min) return min;
-        if(x > max) return max;
-        return x;
-    }
-
-    // str byteToHex(uint8 byte)
-    // converts a single byte to a hex string
-    static byteToHex(byte : byte) : string {
-        return ('0' + byte.toString(16)).slice(-2);
-    }
-
-    // str generateId(int len);
-    // len - must be an even number (default: 32)
-    static generateID() {
-        const len = 32;
-        let arr = new Uint8Array(len / 2);
-        window.crypto.getRandomValues(arr);
-        return Array.from(arr, Helpers.byteToHex).join('');
-    }
-
-    static distance(p : IVector, q : IVector) : float {
-        let dx   = p.x - q.x;
-        let dy   = p.y - q.y;
-        let dz   = p.z - q.z;
-        let dist = Math.sqrt(dx * dx + dy * dy + dz * dz);
-        return dist;
-    }
-
-    // getRandomInt возвращает случайное целое число в диапазоне от min до max (min <= N <= max)
-    static getRandomInt(min : number, max : number) : float {
-        min = Math.ceil(min);
-        max = Math.floor(max);
-        return Math.floor(Math.random() * (max - min + 1)) + min;
-    }
-
-    static createSkinLayer2(text : string, image : HTMLImageElement, callback: Function) {
-        const canvas          = document.createElement('canvas');
-        canvas.width        = 64;
-        canvas.height       = 64;
-        const ctx             = canvas.getContext('2d');
-        if(!ctx) {
-            throw 'error_empty_drawing_context'
-        }
-        if(text) {
-            ctx.fillStyle       = '#f5f5f5';
-            ctx.fillRect(0, 0, 200, 200);
-            ctx.font            = 'bold 20px Arial';
-            ctx.fillStyle       = '#333333';
-            ctx.textAlign       = 'start';
-            ctx.textBaseline    = 'top';
-            ctx.fillText(text, 10, 10);
-        } else {
-            // img, sx, sy, swidth, sheight, x, y, width, height
-            // head
-            ctx.drawImage(image, 32, 0, 32, 16, 0, 0, 32, 16);
-            // body + right leg + right arm
-            ctx.drawImage(image, 0, 32, 56, 16, 0, 16, 56, 16);
-            // left leg
-            ctx.drawImage(image, 0, 48, 16, 16, 16, 48, 16, 16);
-            // left arm
-            ctx.drawImage(image, 0, 48, 48, 16, 32, 48, 16, 16);
-        }
-        // Debug
-        // var link = document.createElement('a');
-        // link.download = 'filename.png';
-        // link.href = canvas.toDataURL()
-        // link.click();
-        canvas.toBlob(function(blob) {
-            let filefromblob = new File([blob], 'image.png', {type: 'image/png'});
-            callback(filefromblob);
-        }, 'image/png');
-    }
-
-    // Canvas download
-    static downloadBlobPNG(blob : Blob, filename : string) {
-        /// create an "off-screen" anchor tag
-        let lnk : HTMLAnchorElement = document.createElement('a'), event;
-        /// the key here is to set the download attribute of the a tag
-        lnk.download = filename;
-        /// convert canvas content to data-uri for link. When download
-        /// attribute is set the content pointed to by link will be
-        /// pushed as "download" in HTML5 capable browsers
-        lnk.href = URL.createObjectURL(blob);
-        /// create a "fake" click-event to trigger the download
-        if (document.createEvent) {
-            event = document.createEvent('MouseEvents');
-            event.initMouseEvent('click', true, true, window,
-                0, 0, 0, 0, 0, false, false, false,
-                false, 0, null);
-            lnk.dispatchEvent(event);
-        } else {
-            (lnk as any).fireEvent?.('onclick');
-        }
-    }
-
-    // downloadImage
-    static downloadImage(image : HTMLImageElement, filename : string) {
-        var c = document.createElement('canvas');
-        var ctx = c.getContext('2d');
-        if(!ctx) {
-            throw 'error_empty_ctx'
-        }
-        ctx.canvas.width  = image.width;
-        ctx.canvas.height = image.height;
-        ctx.drawImage(image, 0, 0);
-        c.toBlob(function(blob) {
-            // here the image is a blob
-            Helpers.downloadBlobPNG(blob, filename);
-        }, 'image/png');
-    }
-
-    static deg2rad(degrees : float) : float {
-        return degrees * (Math.PI / 180);
-    }
-
-    static rad2deg(radians : float) : float {
-        return radians * 180 / Math.PI;
-    }
-
-    static async loadJSON(url : string, callback : Function) {
-        loadText(url, function(text) {
-            callback(JSON.parse(text));
-        });
-    }
-
-    // createGLProgram...
-    static createGLProgram(gl, obj, callback?) {
-        let program = gl.createProgram();
-        // Compile vertex shader
-        let vertexShader = gl.createShader(gl.VERTEX_SHADER);
-        gl.shaderSource(vertexShader, obj.vertex);
-        gl.compileShader(vertexShader);
-        gl.attachShader(program, vertexShader);
-        gl.deleteShader(vertexShader);
-        if(!gl.getShaderParameter(vertexShader, gl.COMPILE_STATUS)) {
-            throw "Could not compile vertex shader!\n" + gl.getShaderInfoLog(vertexShader);
-        }
-        // Compile fragment shader
-        let fragmentShader = gl.createShader(gl.FRAGMENT_SHADER);
-        gl.shaderSource(fragmentShader, obj.fragment);
-        gl.compileShader(fragmentShader);
-        gl.attachShader(program, fragmentShader);
-        gl.deleteShader(fragmentShader);
-        if(!gl.getShaderParameter(fragmentShader, gl.COMPILE_STATUS)) {
-            throw "Could not compile fragment shader!\n" + gl.getShaderInfoLog(fragmentShader);
-        }
-        // Finish program
-        gl.linkProgram(program);
-        if(!gl.getProgramParameter(program, gl.LINK_STATUS)) {
-            throw 'Could not link the shader program!';
-        }
-
-        callback && callback({
-            program
-        });
-
-        return program;
-    }
-
-    // Return from green to red color depend on percentage
-    static getColorForPercentage(pct : float) : Color {
-        var percentColors = [
-            {pct: 0.0, color: {r: 0xff, g: 0x00, b: 0}},
-            {pct: 0.5, color: {r: 0xff, g: 0xff, b: 0}},
-            {pct: 1.0, color: {r: 0x00, g: 0xff, b: 0}}
-        ];
-        for (var i = 1; i < percentColors.length - 1; i++) {
-            if (pct < percentColors[i].pct) {
-                break;
-            }
-        }
-        var lower = percentColors[i - 1];
-        var upper = percentColors[i];
-        var range = upper.pct - lower.pct;
-        var rangePct = (pct - lower.pct) / range;
-        var pctLower = 1 - rangePct;
-        var pctUpper = rangePct;
-        var color = {
-            r: Math.floor(lower.color.r * pctLower + upper.color.r * pctUpper),
-            g: Math.floor(lower.color.g * pctLower + upper.color.g * pctUpper),
-            b: Math.floor(lower.color.b * pctLower + upper.color.b * pctUpper)
-        };
-        return new Color(color.r, color.g, color.b, 1);
-        // or output as hex if preferred
-    }
-
-    // Return speed
-    static calcSpeed(pos1 : Vector, pos2 : IVector, delta : float) : float {
-        return Math.round(pos1.distance(pos2) / delta * 360) / 100;
-    }
-}
-
-// Make fetch functions
-if(typeof fetch === 'undefined') {
-    // Hello eval ;)
-    const code = `Helpers.fetch = (url) => import(url);
-    Helpers.fetchJSON = (url) => import(url, {assert: {type: 'json'}}).then(response => response.default);
-    Helpers.fetchBinary = async (url) => {
-        let binary = fs.readFileSync(url);
-        return binary.buffer;
-    };`;
-    var obj = Helpers;
-    var func = new Function("Helpers", "window", "'use strict';" + code);
-    func.call(obj, obj, obj);
-} else {
-    Helpers.fetch = async (url : string) => fetch(url);
-    Helpers.fetchJSON = async (url, useCache = false, namespace = '') => {
-        const cacheKey = namespace + '|' + url;
-
-        if (useCache && Helpers.cache.has(cacheKey)) {
-            return Promise.resolve(JSON.parse(Helpers.cache.get(cacheKey)));
-        }
-
-        const respt = await fetch(url);
-
-        // if cache is presented - store text response
-        // then we can use this inside a worker
-        if (useCache) {
-            const text = await respt.text();
-
-            Helpers.cache.set(cacheKey, text);
-
-            return JSON.parse(text);
-        }
-
-        return respt.json()
-    };
-
-    Helpers.fetchBinary = async (url) => fetch(url).then(response => response.arrayBuffer());
-}
