+++ www/src/helpers/index.ts
@@ -4,7 +4,6 @@ export * from "./mth.js";
 export * from './vector.js';
 export * from './vector_collector.js';
 export * from './performance_timer.js';
-export * from './helpers.js';
 export * from './string_helpers.js';
 export * from './object_helpers.js';
 export * from './alphabet_texture.js';
@@ -21,3 +20,4 @@ export * from "./array_or_map.js";
 export * from "./array_or_scalar.js";
 export * from "./spatial_determenistic_random.js";
 export * from "./average_clock_timer.js";
+export * from "./simple_shifted_3d_array.js";
